Declare deaths as double

+++ src/main/java/org/csgames/spaceship/sdk/accept/userstory/UserStory.java
@@ -8,9 +8,9 @@ public class UserStory {
   public final List<Scenario> scenarios;
   public final int points;
   public final int penalties;
-  public final int deaths;
+  public final double deaths;
 
-  public UserStory(String name, List<Scenario> scenarios, int points, int penalties, int deaths) {
+  public UserStory(String name, List<Scenario> scenarios, int points, int penalties, double deaths) {
     this.name = name;
     this.scenarios = scenarios;
     this.points = points;
